fix: fs-rename add check for the existance of a new name

+++ src/fs/rename.js
@@ -1,6 +1,11 @@
 import fs from "node:fs/promises";
+import { existsSync } from "node:fs";
 
 const rename = async () => {
+  if (existsSync("./src/fs/files/properFileName.md")) {
+    throw Error("FS operation failed");
+  }
+
   fs.rename(
     "./src/fs/files/wrongFilename.txt",
     "./src/fs/files/properFileName.md"
